refactoring into components before merge
